good progress on moving this stuff over
