+++ ogma/src/lang/help.rs
@@ -74,45 +74,45 @@ pub struct HelpExample {
 mod tests {
     use super::*;
 
-#[test]
-fn help_msg() {
-    use HelpParameter::*;
+    #[test]
+    fn help_msg() {
+        use HelpParameter::*;
 
-    let h = HelpMessage {
-        desc: "this is a description".into(),
-        params: vec![Required("required1".into()), Required("req2".into())],
-        ..HelpMessage::new("cmd-name")
-    };
-    let s = help_as_error(&h).to_string();
-    assert_eq!(
-        &s,
-        "Help: `cmd-name`
+        let h = HelpMessage {
+            desc: "this is a description".into(),
+            params: vec![Required("required1".into()), Required("req2".into())],
+            ..HelpMessage::new("cmd-name")
+        };
+        let s = help_as_error(&h).to_string();
+        assert_eq!(
+            &s,
+            "Help: `cmd-name`
 --> shell:0
  | this is a description
  | 
  | Usage:
  |  => cmd-name required1 req2
 "
-    );
+        );
 
-    let h = HelpMessage {
-        desc: "this is a description".into(),
-        examples: vec![
-            HelpExample {
-                desc: "example 1",
-                code: "cmd-name this is a thingo",
-            },
-            HelpExample {
-                desc: "example 2",
-                code: "cmd-name ",
-            },
-        ],
-        ..HelpMessage::new("cmd-name")
-    };
-    let s = help_as_error(&h).to_string();
-    assert_eq!(
-        &s,
-        "Help: `cmd-name`
+        let h = HelpMessage {
+            desc: "this is a description".into(),
+            examples: vec![
+                HelpExample {
+                    desc: "example 1",
+                    code: "cmd-name this is a thingo",
+                },
+                HelpExample {
+                    desc: "example 2",
+                    code: "cmd-name ",
+                },
+            ],
+            ..HelpMessage::new("cmd-name")
+        };
+        let s = help_as_error(&h).to_string();
+        assert_eq!(
+            &s,
+            "Help: `cmd-name`
 --> shell:0
  | this is a description
  | 
@@ -126,7 +126,6 @@ fn help_msg() {
  |  example 2
  |  => cmd-name 
 "
-    );
+        );
+    }
 }
-}
-
