+++ ogma/tests/commands/io.rs
@@ -95,4 +95,3 @@ fn ls_err() {
         );
     }
 }
-
