+++ ogma/tests/commands/pipeline.rs
@@ -925,4 +925,3 @@ fn to_str_testing() {
     let x = process_w_num("to-str", defs);
     assert_eq!(x, Ok(Value::Str("3.0".into())));
 }
-
